time-data component
