+++ src/asgard_time_advance.hpp
@@ -1,8 +1,5 @@
 #pragma once
-#include "asgard_reconstruct.hpp"
-#include "asgard_transformations.hpp"
-#include "asgard_moment.hpp"
-#include "asgard_solver.hpp"
+#include "asgard_time_data.hpp"
 
 /*!
  * \internal
